Changes to World Creation Page
This originally came from Cameron who is having issues Committing right now. He was having issues with the carousel. I took a look at it and got it working (partially). looks like there may need to be some tweaking with CSS.  Or, at this stage, my personal thought is to use the updated menu at the bottom (would then add in the corresponding sections below - so more coding needed, but mainly stealing from the main landing page and re-working as needed).