finishing character creation
put together the last 3 tertiary pages: leveling, character sheets, and the example characters. Also fiddled with download links for the blank and filled out character sheets and the ability to click on the characters for an expanded view. Additionally changed the character image on the home page to remove the learn more tag. If anyone knows how to do a close button for the image views I have set up that would be the last thing I'd like to add if anything.